// src/generated/models/Cgmp_changesModel.ts
export const Cgmp_changescgmp_category = {
  100000000: 'Network',
  100000001: 'Citrix',
  100000002: 'VMware',
  100000003: 'Firewall',
  100000004: 'Patch',
  100000005: 'Application',
  100000006: 'Database',
  100000007: 'Security',
  100000008: 'Infrastructure',
  100000009: 'Other'
} as const;
export type Cgmp_changescgmp_category = keyof typeof Cgmp_changescgmp_category;
export const Cgmp_changescgmp_changetype = {
  100000000: 'Standard',
  100000001: 'Emergency',
  100000002: 'Normal',
  100000003: 'Expedited'
} as const;
export type Cgmp_changescgmp_changetype = keyof typeof Cgmp_changescgmp_changetype;
export const Cgmp_changescgmp_impactlevel = {
  100000000: 'Low',
  100000001: 'Medium',
  100000002: 'High',
  100000003: 'Critical'
} as const;
export type Cgmp_changescgmp_impactlevel = keyof typeof Cgmp_changescgmp_impactlevel;
export const Cgmp_changescgmp_risklevel = {
  100000000: 'Low',
  100000001: 'Medium',
  100000002: 'High',
  100000003: 'Critical'
} as const;
export type Cgmp_changescgmp_risklevel = keyof typeof Cgmp_changescgmp_risklevel;
export const Cgmp_changescgmp_status = {
  100000000: 'Draft',
  100000001: 'Published',
  100000002: 'UnderReview',
  100000003: 'Released',
  100000004: 'UATUpdates',
  100000005: 'Locked',
  100000006: 'InProgress',
  100000007: 'Completed',
  100000008: 'Failed',
  100000009: 'Closed',
  100000010: 'Cancelled'
} as const;
export type Cgmp_changescgmp_status = keyof typeof Cgmp_changescgmp_status;
export const Cgmp_changescgmp_pirstatus = {
  100000000: 'Draft',
  100000001: 'Submitted',
  100000002: 'Approved',
  100000003: 'Rejected'
} as const;
export type Cgmp_changescgmp_pirstatus = keyof typeof Cgmp_changescgmp_pirstatus;
export const Cgmp_changesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_changesstatecode = keyof typeof Cgmp_changesstatecode;
export const Cgmp_changesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_changesstatuscode = keyof typeof Cgmp_changesstatuscode;

export interface Cgmp_changesBase {
  cgmp_actualendtime?: string;
  cgmp_actualstarttime?: string;
  cgmp_attachmentids?: string;
  cgmp_category: Cgmp_changescgmp_category;
  cgmp_changeid: string;
  cgmp_changenumber: string;
  cgmp_changepoc?: string;
  cgmp_changetype: Cgmp_changescgmp_changetype;
  cgmp_closureremarks?: string;
  cgmp_country?: string;
  cgmp_createdby?: string;
  cgmp_description?: string;
  cgmp_downtimeminutes?: number;
  cgmp_endtime: string;
  cgmp_facility?: string;
  cgmp_impactlevel: Cgmp_changescgmp_impactlevel;
  cgmp_isdeleted?: boolean;
  cgmp_isemergency?: boolean;
  cgmp_ismsignoffat?: string;
  cgmp_ismsignoffby?: string;
  cgmp_isweekend?: boolean;
  cgmp_lessonslearned?: string;
  cgmp_location?: string;
  cgmp_parentchangeid?: string;
  cgmp_pirnotes?: string;
  cgmp_pirstatus?: Cgmp_changescgmp_pirstatus;
  cgmp_projectids?: string;
  cgmp_region?: string;
  cgmp_releasedat?: string;
  cgmp_releasedby?: string;
  cgmp_reviewedat?: string;
  cgmp_reviewedby?: string;
  cgmp_reviewnotes?: string;
  cgmp_risklevel: Cgmp_changescgmp_risklevel;
  cgmp_rollbackinitiated?: boolean;
  cgmp_starttime: string;
  cgmp_status: Cgmp_changescgmp_status;
  cgmp_timeline?: string;
  cgmp_title: string;
  cgmp_uatrequired?: boolean;
  cgmp_uatusers?: string;
  cgmp_versionhistory?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_changesstatecode;
  statuscode?: Cgmp_changesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_changes extends Cgmp_changesBase {
  cgmp_categoryname?: string;
  cgmp_changetypename?: string;
  cgmp_impactlevelname?: string;
  cgmp_pirstatusname?: string;
  cgmp_isdeletedname?: string;
  cgmp_isemergencyname?: string;
  cgmp_isweekendname?: string;
  cgmp_risklevelname?: string;
  cgmp_rollbackinitiatedname?: string;
  cgmp_statusname?: string;
  cgmp_uatrequiredname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}